Users and authentication

+++ src/controllers/expenseController.js
@@ -3,8 +3,10 @@ const db = require('../config/db');
 // GET all expenses
 const getExpenses = async (req, res) => {
   try {
+    const user_id = req.user && req.user.id;
     const result = await db.query(
-      "SELECT * FROM expenses ORDER BY expense_date DESC, id DESC"
+      "SELECT * FROM expenses WHERE user_id = $1 ORDER BY expense_date DESC, id DESC",
+      [user_id]
     );
     res.json(result.rows);
   } catch (err) {
@@ -19,14 +21,16 @@ const getExpenses = async (req, res) => {
 // CREATE expense
 const createExpense = async (req, res) => {
   try {
+
+    const user_id = req.user && req.user.id;
     const { category, amount, expense_date, description } = req.body;
 
     const result = await db.query(
       `INSERT INTO expenses
-      (category, amount, expense_date, description)
-      VALUES ($1, $2, $3, $4)
+      (user_id, category, amount, expense_date, description)
+      VALUES ($1, $2, $3, $4, $5)
       RETURNING *`,
-      [category, amount, expense_date, description]
+      [user_id, category, amount, expense_date, description]
     );
 
     res.json(result.rows[0]);
@@ -42,15 +46,16 @@ const createExpense = async (req, res) => {
 // UPDATE expense
 const updateExpense = async (req, res) => {
   try {
+    const user_id = req.user && req.user.id;
     const { id } = req.params;
     const { category, amount, expense_date, description } = req.body;
 
     const result = await db.query(
       `UPDATE expenses
        SET category = $1, amount = $2, expense_date = $3, description = $4
-       WHERE id = $5
+       WHERE id = $5 AND user_id = $6
        RETURNING *`,
-      [category, amount, expense_date, description, id]
+      [category, amount, expense_date, description, id, user_id]
     );
 
     if (result.rowCount === 0) {
@@ -70,8 +75,9 @@ const updateExpense = async (req, res) => {
 // DELETE expense
 const deleteExpense = async (req, res) => {
   try {
+    const user_id = req.user && req.user.id;
     const { id } = req.params;
-    const result = await db.query("DELETE FROM expenses WHERE id = $1 RETURNING *", [id]);
+    const result = await db.query("DELETE FROM expenses WHERE id = $1 AND user_id = $2 RETURNING *", [id, user_id]);
 
     if (result.rowCount === 0) {
       return res.status(404).json({ message: "Expense not found" });
